moved route parse logic to its own class
