+++ src/Collector.php
@@ -19,6 +19,10 @@ use Codeburner\Router\Exceptions\MethodNotSupportedException;
  * because the routes will not be used until the collector is used.
  */
 
+if (!class_exists(Parser::class, false)) {
+    include __DIR__ . "/Parser.php";
+}
+
 include __DIR__ . "/Route.php";
 include __DIR__ . "/Group.php";
 include __DIR__ . "/Collectors/ControllerCollectorTrait.php";
@@ -35,14 +39,6 @@ class Collector
 
     use Collectors\ControllerCollectorTrait;
     use Collectors\ResourceCollectorTrait;
-
-    /**
-     * These regex define the structure of a dynamic segment in a pattern.
-     *
-     * @var string
-     */
-
-    const DYNAMIC_REGEX = "{\s*(\w*)\s*(?::\s*([^{}]*(?:{(?-1)}*)*))?\s*}";
 
 
     /**
@@ -73,24 +69,23 @@ class Collector
     protected $dynamics = [];
 
     /**
-     * Some regex wildcards for easily definition of dynamic routes. ps. all keys and values must start with :
+     * The pattern parser instance.
      *
-     * @var array
+     * @var Parser
      */
 
-    protected $wildcards = [
-        ":uid"     => ":uid-[a-zA-Z0-9]",
-        ":slug"    => ":[a-z0-9-]",
-        ":string"  => ":\w",
-        ":int"     => ":\d",
-        ":integer" => ":\d",
-        ":float"   => ":[-+]?\d*?[.]?\d",
-        ":double"  => ":[-+]?\d*?[.]?\d",
-        ":hex"     => ":0[xX][0-9a-fA-F]",
-        ":octal"   => ":0[1-7][0-7]",
-        ":bool"    => ":1|0|true|false|yes|no",
-        ":boolean" => ":1|0|true|false|yes|no",
-    ];
+    protected $parser;
+
+    /**
+     * Collector constructor.
+     *
+     * @param Parser|null $parser
+     */
+
+    public function __construct(Parser $parser = null)
+    {
+        $this->parser = $parser ?: new Parser;
+    }
 
     /**
      * @param string $method
@@ -105,8 +100,8 @@ class Collector
 
     public function set($method, $pattern, $action)
     {
-        $method   = $this->parseMethod($method);
-        $patterns = $this->parsePattern($pattern);
+        $method   = $this->getValidMethod($method);
+        $patterns = $this->parser->parsePattern($pattern);
         $group    = new Group;
 
         foreach ($patterns as $pattern)
@@ -206,87 +201,6 @@ class Collector
     }
 
     /**
-     * Determine if the http method is valid.
-     *
-     * @param string $method
-     * @throws MethodNotSupportedException
-     * @return string
-     */
-
-    protected function parseMethod($method)
-    {
-        $method = strtolower($method);
-
-        if (strpos(self::HTTP_METHODS, $method) === false) {
-            throw new MethodNotSupportedException($method);
-        }
-
-        return $method;
-    }
-
-    /**
-     * Separate routes pattern with optional parts into n new patterns.
-     *
-     * @param string $pattern
-     * @return array
-     */
-
-    protected function parsePattern($pattern)
-    {
-        $withoutClosing = rtrim($pattern, "]");
-        $closingNumber  = strlen($pattern) - strlen($withoutClosing);
-
-        $segments = preg_split("~" . self::DYNAMIC_REGEX . "(*SKIP)(*F)|\[~x", $withoutClosing);
-        $this->parseSegments($segments, $closingNumber, $withoutClosing);
-
-        return $this->buildSegments($segments);
-    }
-
-    /**
-     * Parse all the possible patterns seeking for an incorrect or incompatible pattern.
-     *
-     * @param string[] $segments       Segments are all the possible patterns made on top of a pattern with optional segments.
-     * @param int      $closingNumber  The count of optional segments.
-     * @param string   $withoutClosing The pattern without the closing token of an optional segment. aka: ]
-     *
-     * @throws BadRouteException
-     */
-
-    protected function parseSegments(array $segments, $closingNumber, $withoutClosing)
-    {
-        if ($closingNumber !== count($segments) - 1) {
-            if (preg_match("~" . self::DYNAMIC_REGEX . "(*SKIP)(*F)|\]~x", $withoutClosing)) {
-                   throw new BadRouteException(BadRouteException::OPTIONAL_SEGMENTS_ON_MIDDLE);
-            } else throw new BadRouteException(BadRouteException::UNCLOSED_OPTIONAL_SEGMENTS);
-        }
-    }
-
-    /**
-     * @param string[] $segments
-     *
-     * @throws BadRouteException
-     * @return array
-     */
-
-    protected function buildSegments(array $segments)
-    {
-        $pattern  = "";
-        $patterns = [];
-        $wildcardTokens = array_keys($this->wildcards);
-        $wildcardRegex  = $this->wildcards;
-
-        foreach ($segments as $n => $segment) {
-            if ($segment === "" && $n !== 0) {
-                throw new BadRouteException(BadRouteException::EMPTY_OPTIONAL_PARTS);
-            }
-
-            $patterns[] = $pattern .= str_replace($wildcardTokens, $wildcardRegex, $segment);
-        }
-
-        return $patterns;
-    }
-
-    /**
      * @param string $method
      * @param string $pattern
      *
@@ -327,15 +241,32 @@ class Collector
     }
 
     /**
+     * Determine if the http method is valid.
+     *
+     * @param string $method
+     *
+     * @throws MethodNotSupportedException
+     * @return string
+     */
+
+    protected function getValidMethod($method)
+    {
+        $method = strtolower($method);
+
+        if (strpos(self::HTTP_METHODS, $method) === false) {
+            throw new MethodNotSupportedException($method);
+        }
+
+        return $method;
+    }
+
+    /**
      * @return string[]
      */
 
     public function getWildcards()
     {
-        $wildcards = [];
-        foreach ($this->wildcards as $token => $regex)
-            $wildcards[substr($token, 1)] = substr($regex, 1);
-        return $wildcards;
+        return $this->parser->getWildcards();
     }
 
     /**
@@ -344,7 +275,7 @@ class Collector
 
     public function getWildcardTokens()
     {
-        return $this->wildcards;
+        return $this->parser->getWildcardTokens();
     }
 
     /**
@@ -354,7 +285,7 @@ class Collector
 
     public function getWildcard($wildcard)
     {
-        return isset($this->wildcards[":$wildcard"]) ? substr($this->wildcards[":$wildcard"], 1) : null;
+        return $this->parser->getWildcard($wildcard);
     }
 
     /**
@@ -366,7 +297,33 @@ class Collector
 
     public function setWildcard($wildcard, $pattern)
     {
-        $this->wildcards[":$wildcard"] = ":$pattern";
+        $this->parser->setWildcard($wildcard, $pattern);
+        return $this;
+    }
+
+    /**
+     * @return Parser
+     */
+
+    public function getParser()
+    {
+        return $this->parser;
+    }
+
+    /**
+     * @param Parser $parser
+     *
+     * @throws \LogicException
+     * @return self
+     */
+
+    public function setParser(Parser $parser)
+    {
+        if (!empty($this->statics) || !empty($this->dynamics)) {
+            throw new \LogicException("You can't define a route parser after registering a route.");
+        }
+
+        $this->parser = $parser;
         return $this;
     }
